Refactor: Passenger Form moved to Modal
Summary: Removed Passenger screen and moved form content to modal according to actual designs.

Note: This structure will also save us little bit of trouble with passing passenger info between screens. And it will be easier to refactor web version to something closer to kiwi.com later (with form directly on page).
