+++ apps/core/index.js
@@ -2,7 +2,6 @@
 
 export { default as ResultDetail } from './scenes/resultDetail/ResultDetail';
 export { default as Payment } from './scenes/payment/Payment';
-export { default as PassengerForm } from './scenes/passengerForm/PassengerForm';
 export { default as Results } from './scenes/results/Results';
 export { default as Search } from './scenes/search/Search';
 export { default as Bookings } from './scenes/bookings/Bookings';
