+++ apps/core/scenes/resultDetail/ResultDetail.js
@@ -2,7 +2,6 @@
 
 import * as React from 'react';
 import {
-  PassengerCards,
   ContactDetailsForm,
   TableRow,
   TableRowDivider,
@@ -16,6 +15,7 @@ import {
 
 import { PriceSummary } from '../../components/priceSummary';
 import ResultDetailItineraryRenderer from './ResultDetailItineraryRenderer';
+import ResultDetailPassenger from './ResultDetailPassenger';
 
 type Props = {|
   +bookingToken: ?string,
@@ -33,14 +33,6 @@ class ResultDetail extends React.Component<Props, State> {
     email: null,
     phoneNumber: null,
     phoneCountryCode: null,
-  };
-
-  handlePassengerEditPress = id => {
-    if (id) {
-      this.props.navigation.navigate(Routes.PASSENGER_FORM, {
-        id,
-      });
-    }
   };
 
   handleChangeEmail = email => {
@@ -63,49 +55,14 @@ class ResultDetail extends React.Component<Props, State> {
     this.props.navigation.navigate(Routes.PAYMENT);
   };
 
-  handleAddPassenger = () => {
-    // @TODO
-  };
-
   render() {
-    const passengerCards = [
-      {
-        name: 'John Doe',
-        gender: 'male',
-        nationality: 'Russian',
-        dateOfBirth: '22/04/1980',
-        id: 'DF45SV8',
-        insurance: 'Travel Insurance Name',
-        passengerCount: 1,
-        bags: [
-          { count: 2, type: '40x15x30cm, 3kg' },
-          { count: 1, type: '55x20x40cm, 8kg' },
-        ],
-      },
-      {
-        name: 'Jana Nováková',
-        gender: 'female',
-        nationality: 'Czech',
-        dateOfBirth: '22/04/1984',
-        id: 'DF45SV9',
-        insurance: 'Travel Insurance Name',
-        passengerCount: 1,
-        bags: [{ count: 1, type: '40x15x30cm, 3kg' }],
-      },
-    ];
-
     return (
       <>
         <ContentContainer>
           <ResultDetailItineraryRenderer
             bookingToken={this.props.bookingToken}
           />
-          <PassengerCards
-            passengerCards={passengerCards}
-            onActionPress={this.handlePassengerEditPress}
-            actionIconName="edit"
-            onAddPassengerPress={this.handleAddPassenger}
-          />
+          <ResultDetailPassenger />
           <ContactDetailsForm
             onChangeEmail={this.handleChangeEmail}
             onChangePhoneNumber={this.handleChangePhoneNumber}
