--- apps/mobile/src/screens/PassengerFormScreen.js
@@ -1,8 +1,0 @@
-// @flow
-
-import * as React from 'react';
-import { PassengerForm } from '@kiwicom/margarita-core';
-
-export default function PassengerFormScreen() {
-  return <PassengerForm />;
-}
