--- apps/web/pages/passenger-form.js
@@ -1,14 +1,0 @@
-// @flow
-
-import * as React from 'react';
-import { PassengerForm } from '@kiwicom/margarita-core';
-
-import Layout from '../components/Layout';
-
-export default function passengerForm() {
-  return (
-    <Layout>
-      <PassengerForm />
-    </Layout>
-  );
-}
