+++ packages/navigation/config/routes.js
@@ -8,7 +8,6 @@ const Routes = {
   RESULTS: 'results',
   RESULT_DETAIL: 'result-detail',
   PAYMENT: 'payment',
-  PASSENGER_FORM: 'passenger-form',
   BOOKING_LIST: 'booking-list',
   PROFILE: 'profile',
   BOOKING_DETAIL: 'mmb/booking-detail',
